% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggHRUTabletoHUC10.R
\name{aggHRUTableToHUC10}
\alias{aggHRUTableToHUC10}
\title{HRU Table Aggregate to HUC-10}
\usage{
aggHRUTableToHUC10(HRUtable, dateFormat = c("decimal", "string"), dataName,
  dataCategory, timeStep, startDate)
}
\arguments{
\item{HRUtable}{Data.frame of the HRU data to be converted to HUC-10.  Each row = HRU, each column = time step.
Column names should be in either decimal format (produced through lubridate::decimal_date) or date format
(YYYY-MM-DD).  DD is required but irrelevant for time steps monthly or greater.}

\item{dateFormat}{Character; String providing the date format. Either 'decimal' or 'date'.}

\item{dataName}{Character; Name of dataset.}

\item{dataCategory}{Character; Category of dataset (AET, precip, etc).}

\item{timeStep}{Character; Step between measurements (day, month, quarter, annual, etc).}

\item{startDate}{Character; Start date of measurements in 'YYYY-MM-DD' string format.}
}
\value{
Table of HUC-10 means by HUC.  Time-step x HUC-10.
}
\description{
This function aggregates HRU values in table format to HUC-10 values.  It produces the same list output as the 
agg_Raster_to_HUC10 function and output can be fed directly into the downscale_HUC10 function.
Currently requires the following packages: ___.
}
\examples{
aggHRUTableToHUC10()
}
