% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateFeather.R
\name{aggregateFeather}
\alias{aggregateFeather}
\title{Aggregate Time Series Data in Feather Format}
\usage{
aggregateFeather(fNames, fPath = NULL, to = c("month", "year"), fun = sum,
  writeFeather = FALSE, reorder = TRUE, returnOriginals = FALSE)
}
\arguments{
\item{fNames}{Character; Feather filenames to aggregate.}

\item{fPath}{Character; Path to folder containing feather files. If working directory is set to containing
folder, can be left as NULL.}

\item{to}{Character; Target timestep to aggregate to.  Should be either 'month' or 'year'.}

\item{fun}{Function; Function by which to compute summary statistics. Defaults to sum.}

\item{writeFeather}{Logical; Should new feather files be written? Defaults to FALSE.}

\item{reorder}{Logical; Should datasets be reordered by name and HUC level? Defaults to TRUE.}

\item{returnOriginals}{Logical; Return the original feather files in list?}
}
\value{
List of aggregated feather files.  Name of each list item is the updated filename, accessed through names(x),
where x is the returned list.  Original feather files included if returnOriginals == TRUE.
}
\description{
Aggregate already-processed feather files to a lower-resolution time step (eg. daily to monthly).
}
\examples{
aggregateFeather()
}
