% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateRasterToPolygons.R
\name{aggregateRasterToPolygons}
\alias{aggregateRasterToPolygons}
\title{Aggregate Time Series Rasters to Polygons}
\usage{
aggregateRasterToPolygons(dataPath, dataName, dataExtension, dataCategory,
  startDate, timeStep, unitDepth = "mm", aggFUN = "mean",
  MET.HUC10 = TRUE, polyFname = NULL, polyLayer = NULL, polyIDs = NULL,
  maxLayers = 6, cushion = 3, verbose = TRUE, cl = NULL, disag = TRUE,
  cancelReproject = FALSE, recursive = TRUE,
  projManual = "+proj=longlat +datum=NAD83 +no_defs +ellps=GRS80 +towgs84=0,0,0")
}
\arguments{
\item{dataPath}{Character, required; Path to containing folder of gridded datasets.}

\item{dataName}{Character, required; Name of gridded dataset.}

\item{dataExtension}{Character, required; Gridded file extension to import, must be compatible with 'raster' package (tif',
'nc', ect).  The .grd format is discouraged due to connection issues that occur during processing.}

\item{dataCategory}{Character, required; Type of dataset ('precip', 'AET', 'SWE', etc).}

\item{startDate}{Character, required; Starting date of rasters in format "YYYY-MM-DD".  DD required but irrelevant if not daily rasters.}

\item{timeStep}{Character, required; Time step between layers.  Options are 'day', 'week', 'month', 'quarter', and 'year'.}

\item{unitDepth}{Character, required; String providing the unit measurement of depth of dataset.  Converts all datasets to mm.
Defaults to 'mm'.  Supported formats supplied in ?measurements::conv_unit.}

\item{aggFUN}{Character, required; Function to use on cells within each polygon.  Options are c('mean', 'sum').  Use 'mean' 
on rates and 'sum' on volumes. Defaults to 'mean'.}

\item{MET.HUC10}{Logical, optional; If TRUE, uses the included METsteps simplified HUC-10 shapefile.  If FALSE,
requires polyFname, polyLayer, and polyIDs arguments. Defaults to FALSE.}

\item{polyFname}{Character, optional; If MET.HUC10 is FALSE, required filename of shapefile to aggregate gridded dataset. 
Defaults to NULL.}

\item{polyLayer}{Character; If MET.HUC10 is FALSE, required layer of shapefile to aggregate gridded dataset as per 
rgdal::readOGR() function. Defaults to NULL.}

\item{polyIDs}{Character string/vector; If MET.HUC10 is FALSE, required IDs for polygons to be used as rownames in 
return matrix.  Requires either the identification column (in character string) in the shapefile (to be called by 
polys@data[[polyIDs]]) OR a character vector equal in length to the number of features. Defaults to NULL.}

\item{maxLayers}{Integer, optional; Maximum number of layers/bands to allocate to each chunk.  Defaults to 6. If you are
experiencing memory issues (overloading RAM), then drop this number or reduce number of nodes in cl.}

\item{cushion}{Numeric, optional; Number of GB of memory to leave unused during processing.  Strongly advised to not decrease
lower than 2.  Default is 3.}

\item{verbose}{Logical, optional; If FALSE, suppress all text outputs, including the parallel processing progress bar. Defaults 
to TRUE.}

\item{cl}{Cluster object, optional; If supplied, extraction run in parallels.  Suggested max number of cores is 10.}

\item{disag}{Logical; Allow resampling if necessary.  Defaults to TRUE.  Resampling occurs when (average cell size of dataset) x
4 > average size of polygons in shapefile.}

\item{cancelReproject}{Logical; If FALSE, reprojection is forced if necessary.  Set to TRUE to disable raster reprojection.
Defaults to FALSE. Suggested for use when WGS84 is projection and extents align with included datasets (polyHUC2 for example).}

\item{recursive}{Logical; Should the listing recurse into directories? Defaults to TRUE.}

\item{projManual}{Character, optional; Projection for datasets to be converted to.  Gridded data and shapefile projections will be
transformed to projManual as necessary. For CONUS projects, suggested to leave at default. Defaults to NAD83 
('+proj=longlat +datum=NAD83 +no_defs +ellps=GRS80 +towgs84=0,0,0').}
}
\value{
Table of mean raster values for each gridded dataset layer.  Columns are gridded layers, rows are shapefile
features.
}
\description{
Aggregates time-series rasters by polygon primarily through the 'velox' package. Allows the user to take advantage of the
1000x raster extraction speed increases (over 'raster') resulting from the 'velox' package and should be operated in parallels.
Accepts any datatype compatible with the 'raster' package. Defaults to a NAD83 projection and transforms as necessary. Datasets
are assumed to be in depth per measurement and results are calculated as length = mm.
}
\examples{
aggregateRasterToPolygons()
}
