% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arealWeights.R
\docType{data}
\name{arealWeights}
\alias{arealWeights}
\title{HUC 2-10 Areal Weights}
\format{An object of class \code{""list""}; see \code{\link[qtl]{"list"}}.}
\usage{
data(arealWeights)
}
\description{
List of areal weights for HUC levels 2-10, including HUC ID and areal weight (where sum of areal weight for each level
is equal to 1) for each level.
}
\examples{
data(arealWeights)
}
\keyword{datasets}
