% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askRestartR.R
\name{askRestartR}
\alias{askRestartR}
\title{Restart R session question}
\usage{
askRestartR()
}
\value{
If user enters Y, restart R session. Otherwise, return NULL.
}
\description{
Ask user if they would like to restart the current R session.  Accepts 'Y' or 'N', not case sensitive. Used in error
handling when overuse of system memory is identified.
}
\examples{
askRestartR()
}
