% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clRasterize.R
\name{clRasterize}
\alias{clRasterize}
\title{Rasterize Polygons in Parallels}
\usage{
clRasterize(shp, mask.raster, shpValues, cl = NULL, layer = NULL,
  verbose = TRUE, plotMerged = TRUE)
}
\arguments{
\item{shp}{Shapefile to rasterize; Shapefile or path to shapefile.}

\item{mask.raster}{Raster object; NA-filled raster object or path to raster (NA's will be inserted).}

\item{shpValues}{The values to be transferred to raster.  Must be single number or vector of numbers the length of 
the number of spatial features. If vector of numbers, will be assigned in the same order as the order of features in shapefile.}

\item{cl}{Cluster object created through parallel::makeCluster(n). Defaults to NULL, in which case function will not be run
in parallels and will operate more slowly.}

\item{layer}{Layer argument if required for importing shapefile through rgdal::readOGR(). Defaults to NULL. Only required if
shp is path to file.}

\item{verbose}{Show progress? Defaults to TRUE.}

\item{plotMerged}{Plot resulting raster?  Defaults to TRUE, but will change to FALSE if verbose==FALSE.}
}
\value{
features.
}
\description{
Function to chunk features in shapefile, rasterize in parallels, and assign values. Can be performed not in parallels, but is
significantly slower and the user can just as easily use the raster::rasterize() function.
}
\examples{
clRasterize()
}
