% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonalFunctions.R
\name{coin}
\alias{coin}
\title{Coin of data set}
\usage{
coin(x)
}
\arguments{
\item{x}{Vector, matrix, dataframe, or list.}
}
\value{
Head and tail of x.
}
\description{
Returns both the head and tail of dataset
}
