% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMETname.R
\name{createMETname}
\alias{createMETname}
\title{Create File Name}
\usage{
createMETname(x, fromSequence = TRUE)
}
\arguments{
\item{x}{Either list object created by METsteps::downscaleHUC10() function or results from METsteps::extractMetadata() function.}

\item{fromSequence}{Logical; If TRUE, create name from object output from METsteps::downscaleHUC10().  If FALSE, create name
from results of METsteps::extractMetadata().}
}
\value{
List of filenames for zoo objects.
}
\description{
Function to create feather filenames to be called in the METsteps Shiny application.
}
\examples{
createMETname(x)
}
