% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaleHUC10.R
\name{downscaleHUC10}
\alias{downscaleHUC10}
\title{Downscale HUC-10 means to HUCs 2-8}
\usage{
downscaleHUC10(HUC10monthly)
}
\arguments{
\item{HUC10monthly}{Results directly from the METsteps::aggregateRasterToPolygons or METsteps::agg_HRUtable_to_HUC10 function.}
}
\value{
List of tables of HUC 2-8 means in 'zoo' format.  Time-step x HUC-10.
}
\description{
This functions downscales HUC-10 values to the HUC 2-8 levels and, importantly, reformats time-series data into 
zoo objects. Accepts outputs from the METsteps::aggregateRasterToPolygons (when MET.HUC10==T) and the
agg_HRUtable_to_HUC10 functions.
}
\examples{
downscaleHUC10()
}
