% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreShiny.R
\name{exploreShiny}
\alias{exploreShiny}
\title{Shiny Application for Exploring Processed Datasets}
\usage{
exploreShiny(path.feather, path.obsMeta = NULL, path.obs = NULL)
}
\arguments{
\item{path.feather}{Character; Path to folder containing feather files produced through zooHUCtoFeather}

\item{path.obsMeta}{Character; Path to metadata file.  Defaults to NULL.}

\item{path.obs}{Character; Path to folder containing point data observation files as listed in metadata file.}
}
\value{
Shiny application
}
\description{
Launches METsteps Shiny application for interactively exploring datasets, producing statistics, and longer-term, 
calculating a national water budget.
}
\examples{
exploreShiny()
}
