% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{extractMetadata}
\alias{extractMetadata}
\title{Extract Metadata From File Name}
\usage{
extractMetadata(METname, df = TRUE)
}
\arguments{
\item{METname}{List/vector of feather file names.}

\item{df}{Logical; If TRUE, return results as dataframe object. Otherwise, returns list object.  Defaults to TRUE.}
}
\value{
List of filenames for zoo objects.
}
\description{
Function to extract data from feather filenames to be used in METsteps Shiny application
}
\examples{
extractMetadata(zoo.obj)
}
