% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRU_to_HUC10_percData.R
\docType{data}
\name{HRU_to_HUC10_percData}
\alias{HRU_to_HUC10_percData}
\title{Percentages for HRU to HUC10 conversions}
\format{An object of class \code{"list"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(HRU_to_HUC10_percData)
}
\description{
Dataframe of percentages used for conversion between HUCs 2-8 to HUC10 level.
}
\examples{
data(HRU_to_HUC10_percData)
}
\keyword{datasets}
