% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRUids.R
\docType{data}
\name{HRUids}
\alias{HRUids}
\title{List of HRU IDs by Region}
\format{An object of class \code{""list""}; see \code{\link[qtl]{"list"}}.}
\usage{
data(HRUids)
}
\description{
List of HRU IDs by region.
}
\examples{
data(HRUids)
}
\keyword{datasets}
