% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUC10_Shape.R
\docType{data}
\name{HUC10_Shape}
\alias{HUC10_Shape}
\title{HUC-10 Shapefile}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(HUC10_Shape)
}
\description{
Shapefile of the HUC-10 boundaries.  Simplified for use in R through mapshaper.org.
}
\examples{
data(HUC10_Shape)
}
\keyword{datasets}
