% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUC_IDs.R
\docType{data}
\name{HUC_IDs}
\alias{HUC_IDs}
\title{HUC IDs}
\format{An object of class \code{"list"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(HUC_IDs)
}
\description{
List of HUC IDs for HUCs 2-10.  List items are character vectors titled accordingly as HUC2.info, HUC4.info, 
HUC6.info, HUC8.info, and HUC10.info.
}
\examples{
data(HUC_IDs)
}
\keyword{datasets}
