% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUCX_to_HUC10_percData.R
\docType{data}
\name{HUCX_to_HUC10_percData}
\alias{HUCX_to_HUC10_percData}
\title{Percentages for HUCX to HUC10 conversions}
\format{An object of class \code{"list"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(HUCX_to_HUC10_percData)
}
\description{
List of percentages used for conversion between HUCs 2-8 to HUC10 level.
}
\examples{
data(HUCX_to_HUC10_percData)
}
\keyword{datasets}
