% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonalFunctions.R
\name{METclear}
\alias{METclear}
\title{Clear Everything From R Session}
\usage{
METclear(resetR = FALSE)
}
\arguments{
\item{resetR}{Logical; Optional argument to reset entire R session. Helps to clear memory. Defaults to FALSE.}

\item{wd}{Character; Optional character path for new working directory. Defaults to NULL.}
}
\value{
Head of x.
}
\description{
Clears all objects, variables, and figures from R.  Optional argument to reset R.
}
\examples{
METclear()
}
