% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nCell.R
\name{nCell}
\alias{nCell}
\title{Number of Cells in Object}
\usage{
nCell(x)
}
\arguments{
\item{x}{Object; Vector, matrix, dataframe, or list object.}
}
\value{
Number of cells in object.
}
\description{
Function to return total number of cells in objects such as vectors, matrices, dataframes, or lists.
}
