% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteSentence.R
\name{pasteSentence}
\alias{pasteSentence}
\title{Create Properly Structured Sentence}
\usage{
pasteSentence(x, period = FALSE)
}
\arguments{
\item{x}{Character; Vector of strings.}

\item{period}{Logical; If TRUE, return output with concluding period. Defaults to FALSE.}
}
\value{
Shiny application
}
\description{
Paste and collapses vector of strings into correct sentence structure using "," and "and".
}
\examples{
pasteSentence()
}
