% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyHUC10.R
\docType{data}
\name{polyHUC10}
\alias{polyHUC10}
\title{HUC-10 Shapefile}
\format{An object of class \code{""SpatialPolygonsDataFrame""}; see \code{\link[qtl]{"SpatialPolygonsDataFrame"}}.}
\usage{
data(polyHUC10)
}
\description{
Polygon shapefile of simplified HUC-10 delineations.
}
\examples{
data(polyHUC10)
}
\keyword{datasets}
