% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyHUC2.R
\docType{data}
\name{polyHUC2}
\alias{polyHUC2}
\title{HUC-2 Shapefile}
\format{An object of class \code{""SpatialPolygonsDataFrame""}; see \code{\link[qtl]{"SpatialPolygonsDataFrame"}}.}
\usage{
data(polyHUC2)
}
\description{
Polygon shapefile of simplified HUC-2 delineations.
}
\examples{
data(polyHUC2)
}
\keyword{datasets}
