% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyHUC4.R
\docType{data}
\name{polyHUC4}
\alias{polyHUC4}
\title{HUC-4 Shapefile}
\format{An object of class \code{""SpatialPolygonsDataFrame""}; see \code{\link[qtl]{"SpatialPolygonsDataFrame"}}.}
\usage{
data(polyHUC4)
}
\description{
Polygon shapefile of simplified HUC-4 delineations.
}
\examples{
data(polyHUC4)
}
\keyword{datasets}
