% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyHUC6.R
\docType{data}
\name{polyHUC6}
\alias{polyHUC6}
\title{HUC-6 Shapefile}
\format{An object of class \code{""SpatialPolygonsDataFrame""}; see \code{\link[qtl]{"SpatialPolygonsDataFrame"}}.}
\usage{
data(polyHUC6)
}
\description{
Polygon shapefile of simplified HUC-6 delineations.
}
\examples{
data(polyHUC6)
}
\keyword{datasets}
