% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyHUC8.R
\docType{data}
\name{polyHUC8}
\alias{polyHUC8}
\title{HUC-8 Shapefile}
\format{An object of class \code{""SpatialPolygonsDataFrame""}; see \code{\link[qtl]{"SpatialPolygonsDataFrame"}}.}
\usage{
data(polyHUC8)
}
\description{
Polygon shapefile of simplified HUC-8 delineations.
}
\examples{
data(polyHUC8)
}
\keyword{datasets}
