% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmFactors.R
\name{rmFactors}
\alias{rmFactors}
\title{Remove factors}
\usage{
rmFactors(x)
}
\arguments{
\item{x}{Object in vector, matrix, or dataframe format, with uniform type.  To check, use 'typeof' function on object.
Must be 'integer', 'double', or 'character'.}
}
\value{
Input data without factors
}
\description{
This function removes factors from numeric vector, matrix, or dataframe.  Accepts data UNIFORM IN TYPE of types
'integer', 'double', or 'character'.
}
\examples{
rmFactors(x)
}
