% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonalFunctions.R
\name{shead}
\alias{shead}
\title{Short head of object}
\usage{
shead(x, n_col = 1:10, returnDim = TRUE)
}
\arguments{
\item{x}{Vector, matrix, dataframe, or list.}

\item{n_col}{Vector; Which columns to return. Defaults to 10.}

\item{returnDim}{Logical; Return dimensions of data? Defaults to TRUE}
}
\value{
Head of x.
}
\description{
Returns a more limited head (~10 col) of the object.
}
