% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyStats.R
\name{shinyFun_one_90th}
\alias{shinyFun_one_90th}
\title{90th Percentile Value of HUC for selected models}
\usage{
shinyFun_one_90th(x)
}
\value{
Numeric vector.
}
\description{
For use in Shiny app
}
