% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyStats.R
\name{shinyFun_one_SLOPE}
\alias{shinyFun_one_SLOPE}
\title{Lapply version of SLOPE}
\usage{
shinyFun_one_SLOPE(x = subToHUC, alpha = 0.05, insigValue = NA)
}
\value{
Numeric vector.
}
\description{
For use in Shiny app
}
