% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyStats.R
\name{shinyFun_two_Difference}
\alias{shinyFun_two_Difference}
\title{Lapply version of difference}
\usage{
shinyFun_two_Difference(z)
}
\value{
Numeric vector.
}
\description{
For use in Shiny app
}
