% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyStats.R
\name{shinyFun_two_Peak_Timing}
\alias{shinyFun_two_Peak_Timing}
\title{Lapply version of peak function}
\usage{
shinyFun_two_Peak_Timing(x)
}
\value{
Numeric vector.
}
\description{
For use in Shiny app
}
