% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyStats.R
\name{shinyFun_two_Percent_Bias}
\alias{shinyFun_two_Percent_Bias}
\title{Lapply version of PBIAS}
\usage{
shinyFun_two_Percent_Bias(x)
}
\value{
Numeric vector.
}
\description{
For use in Shiny app
}
