% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyStats.R
\name{shinyFun_two_Spearman_Cor}
\alias{shinyFun_two_Spearman_Cor}
\title{Lapply version of spearman correlation}
\usage{
shinyFun_two_Spearman_Cor(z)
}
\value{
Numeric vector.
}
\description{
For use in Shiny app
}
