% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyStats.R
\name{shinyFun_two_Standard_Dev}
\alias{shinyFun_two_Standard_Dev}
\title{Lapply Version of Standard Deviation}
\usage{
shinyFun_two_Standard_Dev(z)
}
\value{
Numeric vector.
}
\description{
For use in Shiny app
}
