% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPlot_HUC_Mean_Percentile_and_ECDF.R
\name{shinyPlot_HUC_Mean_Percentile_and_ECDF}
\alias{shinyPlot_HUC_Mean_Percentile_and_ECDF}
\title{Mean Percentile and ECDF Plots}
\usage{
shinyPlot_HUC_Mean_Percentile_and_ECDF(default. = FALSE,
  subToHUC. = subToHUC, dataCategory. = dataCategory, ablCol = "darkgrey",
  cbPalette. = cbPalette, dnames. = dnames, ...)
}
\arguments{
\item{x}{Numeric; The values to be plotted.}

\item{col}{Character; The plot color.}
}
\value{
Numeric vector.
}
\description{
Description goes here.
}
\examples{
shinyPlot_HUC_Time_Series_and_Difference()
}
