% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPlot_HUC_mvtsplot.R
\name{shinyPlot_HUC_mvtsplot}
\alias{shinyPlot_HUC_mvtsplot}
\title{Multivariate time series plot. Initiated after click}
\usage{
shinyPlot_HUC_mvtsplot(x = subToHUC, ...)
}
\arguments{
\item{x}{Numeric; The values to be plotted..}
}
\value{
.
}
\description{
Description goes here.
}
\examples{
shinyPlot_HUC_mvts(x)
}
