% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPlot_HUC_Seasonal_Decomposition.R
\name{shinyPlot_HUC_Seasonal_Decomposition}
\alias{shinyPlot_HUC_Seasonal_Decomposition}
\title{Decomposition Plot
Decomposes time series into seasonal, irregular, and trend components.
Source: R. B. Cleveland, W. S. Cleveland, J.E. McRae, and I. Terpenning (1990) STL: A Seasonal-Trend Decomposition Procedure Based on Loess. Journal of Official Statistics, 6, 3--73.
Author: Flannery Dolan, Date Created: 07-17-17, Last Edited: 07-18-17}
\usage{
shinyPlot_HUC_Seasonal_Decomposition(default. = FALSE, feederList. = NULL,
  subToHUC. = subToHUC, timeStep. = timeStep, cbPalette. = cbPalette, ...)
}
\arguments{
\item{subToHUC.}{Numeric; .}
}
\value{
Seasonal decomposition plot.
}
\description{
Decomposition Plot
Decomposes time series into seasonal, irregular, and trend components.
Source: R. B. Cleveland, W. S. Cleveland, J.E. McRae, and I. Terpenning (1990) STL: A Seasonal-Trend Decomposition Procedure Based on Loess. Journal of Official Statistics, 6, 3--73.
Author: Flannery Dolan, Date Created: 07-17-17, Last Edited: 07-18-17
}
\examples{
shinyPlot_HUC_Seasonal_Decomposition()
}
