% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPlot_HUC_subHUC_Plot.R
\name{shinyPlot_HUC_subHUC_Plot}
\alias{shinyPlot_HUC_subHUC_Plot}
\title{Time Series of subHUCs Within Selected Region}
\usage{
shinyPlot_HUC_subHUC_Plot(default. = FALSE, feederList. = NULL,
  subToHUC. = subToHUC, subHUC10. = subHUC10,
  logPlotSubset. = logPlotSubset, subsetMonths. = subsetMonths,
  subsetbyMonthsFun. = subsetbyMonthsFun, cbPalette. = cbPalette,
  maphuc. = maphuc, dataCategory. = dataCategory, ablCol = "darkgrey",
  ...)
}
\arguments{
\item{x}{Numeric; The values to be plotted.}

\item{col}{Character; The plot color.}
}
\value{
Numeric vector.
}
\description{
Description goes here.
}
\examples{
shinyPlot_HUC_Time_Series_and_Difference()
}
