% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPlot_HUC_Taylor_Diagram.R
\name{shinyPlot_HUC_Taylor_Diagram}
\alias{shinyPlot_HUC_Taylor_Diagram}
\title{Taylor Diagram}
\usage{
shinyPlot_HUC_Taylor_Diagram(default. = FALSE, subToHUC. = subToHUC,
  cbPalette. = cbPalette, ...)
}
\arguments{
\item{subToHUC.}{Zoo object to plot.}

\item{cbPalette.}{Color values.}
}
\value{
Numeric vector.
}
\description{
Generate Taylor diagram from 'plotrix' package.
}
\examples{
shinyPlot_HUC_Taylor_Diagram()
}
