% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPlot_HUC_Time_Period_Comparison.R
\name{shinyPlot_HUC_Time_Period_Comparison}
\alias{shinyPlot_HUC_Time_Period_Comparison}
\title{Dataset Time Period Comparison}
\usage{
shinyPlot_HUC_Time_Period_Comparison(default. = FALSE, dnames. = dnames,
  maphuc. = maphuc, timeStep. = timeStep, fileInfo. = fileInfo, ...)
}
\arguments{
\item{x}{Numeric; The values to be plotted.}

\item{col}{Character; The plot color.}
}
\value{
Numeric vector.
}
\description{
Plot displaying the temporal overlap of the selected datasets.
}
\examples{
shinyPlot_HUC_Time_Series_and_Difference()
}
