% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPlot_HUC_Time_Series_and_Differences.R
\name{shinyPlot_HUC_Time_Series_and_Difference}
\alias{shinyPlot_HUC_Time_Series_and_Difference}
\title{Primary Time Series and Difference Plot after HUC Click}
\usage{
shinyPlot_HUC_Time_Series_and_Difference(default. = FALSE,
  feederList. = NULL, x = subToHUC, col = cbPalette,
  ablCol = "darkgrey", colDif = "firebrick", HCU. = HCU,
  dataCategory. = dataCategory, multiplot.cex = 1.8, multiplot.lab = 1.8,
  cex.main = 1.8, ...)
}
\arguments{
\item{x}{Numeric; The values to be plotted.}

\item{col}{Character; The plot color.}
}
\value{
Numeric vector.
}
\description{
Description goes here.
}
\examples{
shinyPlot_HUC_Time_Series_and_Difference()
}
