% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPlotExtract.R
\name{shinyPlotExtract}
\alias{shinyPlotExtract}
\title{Return Dataframe of Plotting Functions}
\usage{
shinyPlotExtract(numD = NULL, pkg = "METsteps")
}
\arguments{
\item{numD}{Character, optional; If set to one of c('one', 'two', 'poly'), returns a list of functions for 
that number of input datasets names with formal names ("FunctionName").  If set to NULL, returns dataframe
of all 'shinyFun_' functions. Defaults to NULL.}

\item{pkg}{Character; Which package to look in for function names?  Defaults to "METsteps".}
}
\value{
Numeric vector.
}
\description{
For use in Shiny server. This function returns information (dataframe or list) about each of the plotting
function names.  Function names should be of the format "shinyFun_one_FunctionName",
"shinyFun_two_FunctionName", "shinyFun_poly_FunctionName", where "one", "two", and "poly" tell the app 
how many input datasets the function can handle. If function can handle two or more, copy/paste function 
with new names.  "FunctionName" is the name of the function as it will appear in the Shiny interface. 
Any underscores after c("one", "two", "poly") will be converted to spaces.
}
\examples{
shinyFunExtract()
}
