% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateLines.R
\docType{data}
\name{stateLines}
\alias{stateLines}
\title{CONUS States Shapefile}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(stateLines)
}
\description{
Polyline shapefile of CONUS state delineations.
}
\examples{
data(stateLines)
}
\keyword{datasets}
