% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrRight.R
\name{substrRight}
\alias{substrRight}
\title{Substring of a Character Vector From the Right}
\usage{
substrRight(x, n = 1)
}
\arguments{
\item{x}{Character; Vector}

\item{n}{Integer; Number of characters to keep from the right. Defaults to 1.}
}
\value{
List of filenames for zoo objects.
}
\description{
Extract substrings starting from the right.
}
\examples{
substrRight()
}
