% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemMemory.R
\name{systemMemory}
\alias{systemMemory}
\title{System Memory}
\usage{
systemMemory(units = "GB", returnAsList = TRUE, verbose = FALSE)
}
\arguments{
\item{returnAsList}{Logical; If TRUE, return results as list. If FALSE, return results as named vector. Defaults
to TRUE.}

\item{verbose}{Logical; If TRUE, writes units to console.  Defaults to FALSE.}

\item{dataPath}{Character; Units of memory. Options are c('KB', 'MB', 'GB').  Defaults to 'GB'.}
}
\value{
List or vector of memory use, free memory, and total memory.
}
\description{
Retrieve system memory information using the R base 'system' function to invoke an OS call.  Returns current amount of
memory in use, available free memory, and total visible memory.  Windows Only!
}
\examples{
systemMemory()
}
