% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor.R
\name{taylor}
\alias{taylor}
\title{Taylor Diagram}
\usage{
taylor(allData, dataNames = NULL, dataColors = NULL)
}
\arguments{
\item{dataNames}{Character; Names of datasets in order of c(obs, sim). If NULL, names are defined as 'obs', 'sim1',
'sim2', etc.  Defaults to NULL.}

\item{dataColors}{Character; Specify colors for each dataset.  If NULL, colors are assigned automatically. Defaults
to NULL.}

\item{obs}{Numeric; Standard deviation of observational dataset.}

\item{sim}{Numeric; List of vectors of standard deviations and correlations of each dataset, in that order.}
}
\value{
Taylor diagram
}
\description{
Create Taylor diagram from time-series data.
}
\examples{
taylor()
}
