% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonalFunctions.R
\name{test_match_order}
\alias{test_match_order}
\title{Check matching between vectors}
\usage{
test_match_order(x, y)
}
\arguments{
\item{x}{Vector}

\item{y}{Vector same length as x.}
}
\value{
Logical value of TRUE if x and y are the same.
}
\description{
Returns a more limited head (~10 col) of the object.  There are more options in function that are commented out
that provide more information.
}
