% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeOverlap.R
\name{timeOverlap}
\alias{timeOverlap}
\title{Time Series Intersect}
\usage{
timeOverlap(startDates, endDates, by)
}
\arguments{
\item{startDates}{Character; Start dates. Must be same length as endDates.}

\item{endDates}{Character; End dates. Must be same length as startDates.}

\item{by}{Character; increment of sequence as found in seq.Date() function.  Length of one.}
}
\value{
Numeric vector.
}
\description{
Reduce time series to overlapping time periods with start dates, end dates, and time steps (for continuous datasets).
}
\examples{
timeOverlap()
}
