% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tictoc.R
\name{toc}
\alias{toc}
\title{End tic/toc function}
\usage{
toc(LS = NULL, end = FALSE, return.value = FALSE)
}
\arguments{
\item{LS}{Character; Optional leading string to be included to annotate different timesteps.  Default is NULL.}

\item{end}{Logical; If TRUE, end timing sequence, deletes tic() initialization.  Defaults to FALSE.}

\item{return.value}{Logical; If TRUE, function returns elapsed time value in numeric format for use.  Defaults to FALSE.
Units in seconds.}
}
\value{
Calculate time change since initializing tic function.
}
\description{
Ending function for tic/toc timing sequence.  Requires the tic function to have been run first.
Can be run in series based on original tic function product.
}
