% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterbalanceShiny.R
\name{waterbalanceShiny}
\alias{waterbalanceShiny}
\title{Shiny Application for Calculting CONUS Water Budget}
\usage{
waterbalanceShiny(path.feather)
}
\arguments{
\item{path.feather}{Character; Path to folder containing feather files produced through zooHUCtoFeather}
}
\value{
Shiny application
}
\description{
Launches METsteps Shiny application for interactively calculating a national water balance using feather tables created
within the METsteps workflow.
}
\examples{
waterbalanceShiny()
}
