% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonalFunctions.R
\name{coin}
\alias{coin}
\title{Coin of data set}
\usage{
coin(x, full = FALSE)
}
\arguments{
\item{x}{Vector, matrix, dataframe, or list.}

\item{full}{Logical; Should the full head and full tail be returned instead of shorts?}
}
\value{
Head and tail of x.
}
\description{
Returns both the short head and short tail of dataset
}
