% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMETname.R
\name{createMETname}
\alias{createMETname}
\title{Create File Name}
\usage{
createMETname(zoo.obj)
}
\arguments{
\item{zoo.obj}{List object created by METsteps::downscaleHUC10() function.}
}
\value{
List of filenames for zoo objects.
}
\description{
Function to create feather filenames to be called in the METsteps Shiny application.
}
\examples{
createMETname(zoo.obj)
}
