% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooHUCtoFeather.R
\name{zooHUCtoFeather}
\alias{zooHUCtoFeather}
\title{Convert to feather}
\usage{
zooHUCtoFeather(zoo.obj,
  Shiny.path = "C:/Users/ssaxe/Documents/Projects/Model Evaluation/RShiny/")
}
\arguments{
\item{zoo.obj}{Either output from downscale_HUC10() function (list object of zoo-format dataframes) or
vector of paths to locations of csv files.}

\item{Shiny.path}{Path to save location.  Defaults to current working directory.  Set to your RShiny app directory.
Will create a folder for feather files named exactly 'FeatherFiles' if not already available.}
}
\value{
List of tables of HUC 2-8 means in 'zoo' format.  Time-step x HUC-10.
}
\description{
This function converts the monthly mean HUC values to 'feather' format to enable faster importing than .csv 
or .txt formats.  Necessary for rapid RShiny app processing.  Requires 'feather' package.
}
\examples{
zooHUCtoFeather()
}
